enhance(scripts): add support to build a package from a local source directory set in `$TERMUX_PKG_SRCURL` if its in the format `file:///path/to/source/dir`
The `/path/to/source/dir` must be an absolute and normalized path to a source directory on the local file system.

Whatever was the current state of the local source directory when build was started, that is what will be built, without any changes to any `git` branches/tags. Any uncommitted changes to current `git` branch will also get built. No checksum checks like one set in `$TERMUX_PKG_SHA256` will be done, and a tar for source directory will be created every time package is built, assuming `-f` flags are passed for rebuilds.

The `termux_download()` function has been copied from `master` branch, check it for history of contributions.